:bug: interfaceManager fixed and improved
Rewrote it completely to account for menu shifts due to different
users having different ammounts of Blue Essence and Riot Points.

The interfaceManager now finds the line seperating the menu from
BE & RP, and creates the {x,y} points where each menu item is
located. It uses the color of those pixels to determine in which
client page you are.