:zap: Improved screenshot stitching algorithm
The search algorithm now checks a 160x80px size area for a match
with a high color difference tolerance due to the moving background.
Rather than checking in the usual line by line fashion, it checks
with lines at a 45deg angle to hit a incompatible pixel and disqalify
the match earlier in the process.